pyproject.toml
[project]
name = "ssl-viewer"
version = "0.1.0"
description = "Streamlit SSL viewer for planar-array telemetry"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydeck>=0.9.1",
    "pyzmq>=27.1.0",
    "streamlit>=1.55.0",
]
